+++ src/classes/managers/tower.manager.js
@@ -8,7 +8,7 @@
 //====================================================================================================================
 //====================================================================================================================
 
-import Tower from "../models/tower.model.js";
+import Tower from "../models/tower.class.js";
 
 class TowerManager {
     static instance = null;
